+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @effect/cli
 
+## 0.33.9
+
+### Patch Changes
+
+- Updated dependencies [[`65895ab`](https://github.com/Effect-TS/effect/commit/65895ab982e0917ac92f0827e387e7cf61be1e69)]:
+  - @effect/platform@0.45.1
+
+## 0.33.8
+
+### Patch Changes
+
+- Updated dependencies [[`2b62548`](https://github.com/Effect-TS/effect/commit/2b6254845882f399636d24223c483e5489e3cff4)]:
+  - @effect/platform@0.45.0
+
 ## 0.33.7
 
 ### Patch Changes